requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
Flask-Migrate==4.0.*
Flask-Login==0.6.*
Flask-WTF==1.2.*
SQLAlchemy==2.0.*
psycopg2-binary==2.9.*
pgvector==0.3.*
python-dotenv==1.0.*
Werkzeug==3.1.*
email-validator==2.1.*
sentence-transformers==3.*
gunicorn==22.*
